Change type of iterable variable to int

+++ cpp/square_root/square_root_tests.cpp
@@ -65,7 +65,7 @@ namespace SquareRootTests {
     TEST(SquareRoot, RandomizedTests) {
         std::mt19937 generator(std::chrono::steady_clock::now().time_since_epoch().count());
         std::uniform_int_distribution<unsigned long long> uniform_int_distribution(0LL, ~(((1ULL << 11) | 1) << 52));
-        for (long long i = 0; i < NUMBER_OF_RANDOMIZED_TESTS; ++i) {
+        for (int i = 0; i < NUMBER_OF_RANDOMIZED_TESTS; ++i) {
             unsigned long long random_value = uniform_int_distribution(generator);
             double double_representation = *reinterpret_cast<double *>(&random_value);
             EXPECT_DOUBLE_EQ(SquareRoot(double_representation), sqrt(double_representation));
